split App.vue into sub-components
